@@ -1,6 +1,5 @@
 """Subscribes users to the EFF newsletter."""
 import logging
-from typing import cast
 from typing import Optional
 
 import requests
@@ -33,11 +32,9 @@ def prepare_subscription(config: configuration.NamespaceConfig, acc: Account) ->
         if config.email is None:
             _report_failure("you didn't provide an e-mail address")
         else:
-            # TODO: Remove cast when https://github.com/certbot/certbot/pull/9073 is merged.
-            acc.meta = cast(Account.Meta, acc.meta.update(register_to_eff=config.email))
+            acc.meta = acc.meta.update(register_to_eff=config.email)
     elif config.email and _want_subscription():
-        # TODO: Remove cast when https://github.com/certbot/certbot/pull/9073 is merged.
-        acc.meta = cast(Account.Meta, acc.meta.update(register_to_eff=config.email))
+        acc.meta = acc.meta.update(register_to_eff=config.email)
 
     if acc.meta.register_to_eff:
         storage = AccountFileStorage(config)
@@ -56,11 +53,9 @@ def handle_subscription(config: configuration.NamespaceConfig, acc: Optional[Acc
     if config.dry_run or not acc:
         return
     if acc.meta.register_to_eff:
-        # TODO: Remove cast when https://github.com/certbot/certbot/pull/9073 is merged.
-        subscribe(cast(str, acc.meta.register_to_eff))
+        subscribe(acc.meta.register_to_eff)
 
-        # TODO: Remove cast when https://github.com/certbot/certbot/pull/9073 is merged.
-        acc.meta = cast(Account.Meta, acc.meta.update(register_to_eff=None))
+        acc.meta = acc.meta.update(register_to_eff=None)
         storage = AccountFileStorage(config)
         storage.update_meta(acc)
 
